+++ app/repositories/XiEoriInformationRepository.scala
@@ -24,6 +24,8 @@ import play.api.Configuration
 import play.api.libs.json.{Format, Json, OFormat}
 import uk.gov.hmrc.mongo.MongoComponent
 import uk.gov.hmrc.mongo.play.json.PlayMongoRepository
+import org.mongodb.scala.SingleObservableFuture
+import org.mongodb.scala.ToSingleObservablePublisher
 
 import java.time.LocalDateTime
 import java.util.concurrent.TimeUnit
@@ -31,9 +33,8 @@ import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
 
 @Singleton
-class DefaultXiEoriInformationRepository @Inject()(
-                                                    config: Configuration,
-                                                    mongoComponent: MongoComponent
+class DefaultXiEoriInformationRepository @Inject()(config: Configuration,
+                                                   mongoComponent: MongoComponent
                                                   )(implicit executionContext: ExecutionContext)
   extends PlayMongoRepository[XiEoriInformationMongo](
     collectionName = "xieori-information",
@@ -43,7 +44,7 @@ class DefaultXiEoriInformationRepository @Inject()(
       IndexModel(
         ascending("lastUpdated"),
         IndexOptions().name("xieori-information-last-updated-index")
-          .expireAfter(config.get[Int]("mongodb.timeToLiveInHoursXieoriInformation"), TimeUnit.HOURS)
+          .expireAfter(config.get[Long]("mongodb.timeToLiveInHoursXieoriInformation"), TimeUnit.HOURS)
       ))
   ) with XiEoriInformationRepository {
